backend/requirements.txt
django==5.1.4
djangorestframework==3.15.2
openai==1.57.2
ipython==8.30.0
django-environ==0.11.2